+++ gatsby-plugin-material-ui/.gitignore
@@ -70,3 +70,4 @@ lib
 /gatsby-ssr.js
 /.cache/
 /utils.js
+/autoprefixer.js
