+++ gatsby-plugin-material-ui/src/__tests__/utils.test.js
@@ -1,5 +1,5 @@
-import { hasEntries, prefix } from "../utils";
-import { fail } from "assert";
+import { hasEntries } from "../utils";
+
 describe(`utils`, () => {
   describe(`hasEntries`, () => {
     it(`should return false if object is empty`, () => {
@@ -12,21 +12,6 @@ describe(`utils`, () => {
 
     it(`should return true if object has entries`, () => {
       expect(hasEntries({ jss: {} })).toBe(true);
-    });
-  });
-
-  describe(`prefix`, () => {
-    const invalidCss = `
-        .jss1 {
-            border-radius: 6px;
-        }
-        .jss1 {
-            h1, h2, h3: [object Object];
-        }
-    `;
-
-    it(`should throw an error if css is invalid1`, () => {
-      expect(() => prefix(invalidCss, `/test`)).toThrowErrorMatchingSnapshot();
     });
   });
 });
